+++ changelog.md
@@ -3,6 +3,7 @@
 ## Version 1.12.0 (Future)
 - Make path calculation happen in a queued job
 - Add `loadRelationCanonicalPath` method to entity collection
+- Change back button label on entity list edit form
 
 ## Version 1.11.1 (2017-10-05)
 - Fix canCreateForParent method